Reorganization
Organized some things to make much more sense, started the move to
tick-based game looping. Main game loop will happen in the GameEngine
class, and will call ticks every 1/60 of a second to the
GraphicsWrangler and PhysicsWrangler classes to keep them aligned with
one another. PhysicsWrangler will inform the Engine on whether it is
done doing calculations, and the gameloop will wait for the Physics to
be done calculating.